🗃️ Committing everything that changed 🗃️

+++ etc/functions/python.bash
@@ -10,24 +10,24 @@
 # @@Description      :  activate a python virtual environment
 # @@Changelog        :  newScript
 # @@TODO             :  Refactor code
-# @@Other            :  
-# @@Resource         :  
+# @@Other            :
+# @@Resource         :
 # @@Terminal App     :  no
 # @@sudo/root        :  no
 # @@Template         :  none
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-_venv_completion() { 
+_venv_completion() {
   local venv_name="" venv_dir="" venv_projects=""
   local SETV_VIRTUAL_DIR_PATH="${SETV_VIRTUAL_DIR_PATH:-$VENV_DIR}"
-  venv_name="$(basename "$PWD")"                 
+  venv_name="$(basename "$PWD")"
   venv_dir="${SETV_VIRTUAL_DIR_PATH/$venv_name:-$PWD/.venv}"
-  [ -f "$PWD/.venv_name" ] && venv_dir="$(<"$PWD/.venv_name")" 
-  venv_projects="$([ -n "$SETV_VIRTUAL_DIR_PATH" ] && find "$SETV_VIRTUAL_DIR_PATH/" -maxdepth 1 -type d -printf "%P\n" 2>/dev/null | grep  -v '^$' || echo '')"
+  [ -f "$PWD/.venv_name" ] && venv_dir="$(<"$PWD/.venv_name")"
+  venv_projects="$([ -n "$SETV_VIRTUAL_DIR_PATH" ] && find "$SETV_VIRTUAL_DIR_PATH/" -maxdepth 1 -type d -printf "%P\n" 2>/dev/null | grep -v '^$' || echo '')"
   #[ "$word" = "delete" ] && COMPREPLY=($(compgen -W '${venv_projects}' -- "$cur"))
   if [ -n "$VIRTUAL_ENV" ]; then
     COMPREPLY=($(compgen -W 'deactivate' -- "$cur"))
   elif [ -f "$venv_dir/bin/activate" ]; then
-    COMPREPLY=($(compgen -W 'activate delete"' -- "$cur")) 
+    COMPREPLY=($(compgen -W 'activate delete"' -- "$cur"))
   elif [ -n "$venv_projects" ]; then
     COMPREPLY=($(compgen -W 'create delete ${venv_projects}' -- "$cur"))
   fi
@@ -42,7 +42,7 @@ venv() {
   [ -f "$PWD/.venv_name" ] && venv_dir="$(<"$PWD/.venv_name")"
   builtin type -P python3 &>/dev/null || return 1
   if [ -z "$VIRTUAL_ENV" ]; then
-    if $pythonbin -m virtualenv 2>&1 |& grep -q 'No module named'; then 
+    if $pythonbin -m virtualenv 2>&1 |& grep -q 'No module named'; then
       printf_blue "Attempting to install the virtualenv module"
       eval $pythonbin -m pip install --user virtualenv &>/dev/null || return 1
     fi
@@ -52,19 +52,20 @@ venv() {
     . "$venv_dir/bin/activate"
   elif [ "$1" = "deactivate" ] && [ -n "$VIRTUAL_ENV" ]; then
     deactivate
-  elif [ "$1" = "delete" ] && [ -d "$venv_dir/$2" ]; then
+  elif [ "$1" = "delete" ]; then
+    [ -n "$2" ] || return 1
     printf_purple "Deleting $venv_dir/$2"
-    rm -Rf "$venv_dir/$2"
+    rm -Rf "${venv_dir:?}/${2:-}"
     [ -d "$venv_dir/$2" ] && printf_red "Failed to delete $2" || printf_green "Deleted the environment $2"
   else
     [ -n "$VIRTUAL_ENV" ] && printf_yellow "A virtual environment is already active" && return
-    [ "$1" = "create" ] && shfit 1
-    [ -n "$1" ] && venv_dir="$1" && echo "$venv_dir" > "$PWD/.venv_name"
+    [ "$1" = "create" ] && shift 1
+    [ -n "$1" ] && venv_dir="$1" && echo "$venv_dir" >"$PWD/.venv_name" || { printf_red "Please provide a venv name" && return 1; }
     [ -d "$venv_dir" ] && printf_red "$venv_name already exists" && return 1
     printf_cyan "Creating a new environment for $venv_name"
-    $pythonbin -m venv "$venv_dir" &>/tmp/$venv_name.log && [ -f "$venv_dir/bin/activate" ] && "$venv_dir/bin/activate" && 
+    $pythonbin -m venv "$venv_dir" &>/tmp/$venv_name.log && [ -f "$venv_dir/bin/activate" ] && "$venv_dir/bin/activate" &&
       printf_green "virtual environment has been initiated" ||
-      { printf_red "Failed to initiate the virtual environment" && printf_yellow "in: $venv_dir" && return 1; } 
+      { printf_red "Failed to initiate the virtual environment" && printf_yellow "in: $venv_dir" && return 1; }
   fi
   return 0
 }
@@ -77,7 +78,7 @@ activate() {
       setv "$1" || return 1
     fi
   #else
-  #  type deactivate &>/dev/null && deactivate || unset 
+  #  type deactivate &>/dev/null && deactivate || unset
   fi
   return 0
 } && export -f activate
